add incremental test coverage for tree backup
refactors many operations-level drive-based tests with a focus on
centralization and re-use, then uses those central tests to run
driveish tests using the tree version for all drive-based services.
